12/10commit

+++ GameTemplate/k2EngineLow/graphics/effect/EffectEmitter.cpp
@@ -21,6 +21,11 @@ namespace nsK2EngineLow {
 	{
 		m_effect.Update();
 
+		if (m_isAutoDelete == false)
+		{
+			return;
+		}
+
 		if (!IsPlay())
 		{
 			DeleteGO(this);
